fix(walk): hard links are told apart by each file's exact id, never by a double
The native engine, the NTFS turbo mode and the walker keyed hard links by
file ids converted to doubles (P3-7: "Legacy-identical keys"). A double holds
an id exactly only below 2^53, and an NTFS file reference keeps its record's
sequence number in bits 48..64: a record reused 32 times or more has an id at
or above 2^53, where neighbouring records round to one number —
(40<<48)|4096 and (40<<48)|4097 are both 11258999068430336. On Windows every
readable file is a family candidate (the listing has no link count), so two
unrelated files became one "family": the ingest dropped the second one's
bytes as a duplicate, and refresh_families gave both the first one's size
and times. The listing also kept only the low 64 bits of FILE_ID_128, which
ReFS does not guarantee unique. Found by the pre-landing review's enum and
type-coercion pass, and independently by its API-contract specialist.

- tm-walk: Meta.ino is exact (u128; the whole 128-bit id on Windows). New
  `links` module: hardlink_families groups names by (dev, ino) exactly, by
  sort (no per-candidate allocation, replacing collide_ids' map), and
  HardlinkRef carries a family number instead of dev/ino. tm-mft's tree uses
  the same function, so the two engines cannot drift.
- refresh_families (the reviewers' second finding on it): a read counts only
  when it reaches the family's own file (same dev and id — a folder swapped
  for a link would lead elsewhere; ReFS's 64-bit handle id cannot match its
  128-bit listing id, so such a family keeps its listing's values), checks
  for a cancel before each read and moves the heartbeat, so tens of
  thousands of WinSxS families are neither stuck past a cancel nor taken for
  a stalled walk.
- napi WalkResult: hardlinkDev/hardlinkIno (doubles) → hardlinkFamily
  (Uint32Array); the ingest keys by family number.
- Columns file TMMFT002: hard-link entries are (node, family); a TMMFT001
  file is refused as "written by another TreeMap build"; check_shape refuses
  a family numbered past the table.
- The built-in walker: hardlinkKey reads an id at or past 2^53 once more as a
  bigint (only for shared files, only for those ids); a file gone before
  that second look keeps its first key.
- Docs: P3-7 amended; DESIGN (side table, §9.2 item 3, §16 item 2); RISKS
  R60 records the persistent index's same limit (indexEngine stores ino as
  a double) for Phase 4.

Test-first, each seen RED: the Windows staging test (two ids past 2^53 and
two differing only above bit 64 got one ino), the MFT tree test (records
4096/4097 at sequence 40 made a family), the ingest test (families told
apart by number), the walker key test (b keyed as a's rounded id). Mutants
caught: grouping by double in hardlink_families (tree + walk tests), the
ingest keying every link as one family, the walker's lossy key, each
refresh guard (identity, kind), the refresh's cancel check and its
heartbeat bump, and a TMMFT001 file read as foreign.

Gate: cargo fmt; clippy -D warnings on macOS, x86_64-unknown-linux-gnu and
x86_64-pc-windows-msvc; cargo test --workspace 334/0; npm run typecheck;
npm test 2,937 tests, 2,930 pass, 0 fail, 7 skipped.

+++ native/treemap-core/crates/tm-walk/src/platform/mod.rs
@@ -26,7 +26,7 @@ pub mod windows;
 
 /// Everything the listing knows about one entry besides its name. Times are
 /// milliseconds computed as `sec * 1e3 + nsec / 1e6` without rounding (P3-6);
-/// sizes, `dev` and `ino` are doubles (P3-7).
+/// sizes and `dev` are doubles (P3-7); `ino` is exact, never a double.
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Meta {
     /// [`crate::KIND_FILE`], [`crate::KIND_DIR`] or [`crate::KIND_SYMLINK`].
@@ -43,8 +43,13 @@ pub struct Meta {
     pub atime_ms: f64,
     /// The device id.
     pub dev: f64,
-    /// The inode / file id.
-    pub ino: f64,
+    /// The inode or file id, exact: `st_ino`, or on Windows the whole 128-bit
+    /// file id. An identity, not a quantity — as a double, a reused NTFS
+    /// record's id (its sequence number sits in bits 48..64) rounds into its
+    /// neighbour's past 2^53, and ReFS ids can differ only above bit 64, so
+    /// two files became one hard-link family (the pre-landing review of 23
+    /// Sep 2026).
+    pub ino: u128,
     /// Hard-link count; 0 for a directory (not a fact the walk records) or when withheld.
     pub nlink: u32,
     /// True when the file system withheld an attribute the walk needs; the
@@ -63,7 +68,7 @@ impl Meta {
             mtime_ms: f64::NAN,
             atime_ms: f64::NAN,
             dev: 0.0,
-            ino: 0.0,
+            ino: 0,
             nlink: 0,
             withheld: true,
         }
